requirements.txt
# Core LiveKit agent
livekit-agents[openai,azure,turn-detector]==1.3.3

# Azure SDKs
azure-storage-blob==12.22.0

# OpenAI (Azure OpenAI Client)
openai==1.58.1

# Utils
python-dotenv==1.0.1
